+++ polaris-pipeline/DdeiClient/Factories/IDdeiArgFactory.cs
@@ -16,7 +16,7 @@ namespace Ddei.Factories
             Guid correlationId,
             string urn,
             int caseId,
-            int documentId,
+            long documentId,
             long versionId);
         DdeiCmsDocumentNotesArgDto CreateDocumentNotesArgDto(string cmsAuthValues, Guid correlationId, string urn, int caseId, string documentId);
         DdeiCmsAddDocumentNoteArgDto CreateAddDocumentNoteArgDto(string cmsAuthValues, Guid correlationId, string urn, int caseId, int documentId, string text);
